[#73] refactor: Room entity to access room_id with readable option
