+++ BE/src/main/java/dev/codesquad/airbnb02/domain/favorite/Favorite.java
@@ -1,4 +1,4 @@
-package dev.codesquad.airbnb02.domain.room.data;
+package dev.codesquad.airbnb02.domain.favorite;
 
 import javax.persistence.Column;
 import javax.persistence.Embeddable;
@@ -12,6 +12,9 @@ import lombok.Setter;
 @Setter
 @Embeddable
 public class Favorite {
+
+  @Column(name = "room_id", nullable = false)
+  private Long roomId;
 
   @Column(name = "user_id", nullable = false)
   private Long userId;
